updated a bunch of stuff

+++ kernel/src/terminal/terminal.h
@@ -4,11 +4,11 @@
 
 
 
-#include <stdint.h>
+#include "../superheader.h"
 
 
 void KiTerminalPuts(const char* str);
 void KiTerminalPutChar(char c);
 void KiTerminalClearScreen();
-char* KiTerminalGets(int size);
+char* KiTerminalGets(size_t size);
 #endif
